moved the comparison function according to the todo comment
