+++ src/utils/mlx_list.c
@@ -6,7 +6,7 @@
 /*   By: W2Wizard <w2.wizzard@gmail.com>              +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/12/28 01:53:51 by W2Wizard      #+#    #+#                 */
-/*   Updated: 2022/11/26 14:26:23 by jvan-hal      ########   odam.nl         */
+/*   Updated: 2023/01/31 17:23:11 by jvan-hal      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -90,23 +90,6 @@ void mlx_lstadd_front(mlx_list_t** lst, mlx_list_t* new)
 	new->next = *lst;
 	new->prev = NULL;
 	*lst = new;
-}
-
-// TODO: Move this somewhere else...
-bool mlx_equal_image(void* lstcontent, void* value)
-{
-	const mlx_image_t* lcontent = lstcontent;
-	const mlx_image_t* lvalue = value;
-
-	return (lcontent == lvalue);
-}
-
-bool mlx_equal_inst(void* lstcontent, void* value)
-{
-	const draw_queue_t* lcontent = lstcontent;
-	const mlx_image_t* lvalue = value;
-
-	return (lcontent->image == lvalue);
 }
 
 /**
